fix prefix continuos version

+++ Problemi svolti/Prefix Sum/4-2 C-Index x A(Continuos ver.)/tempCodeRunnerFile.cpp
@@ -6,7 +6,7 @@ int main()
       int N, M;
       cin >> N >> M;
 
-      vector<int64_t> A(N);
+      vector<int> A(N);
       for (int i = 0; i < N; i++)
       {
             cin >> A[i];
